requirements.txt
alabaster==0.7.9
Babel==2.3.4
colorama==0.3.7
coverage==4.2
decorator==4.0.10
Django==1.10
django-bootstrap==0.2.4
django-bootstrap-form==3.2.1
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
docutils==0.12
factory-boy==2.7.0
fake-factory==0.7.2
imagesize==0.7.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
numpy==1.11.1
pandas==0.18.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
Pygments==2.1.3
python-dateutil==2.5.3
python-openexchangerates==0.1.0
pytz==2016.6.1
requests==2.11.1
selenium==2.53.6
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
sqlparse==0.2.1
traitlets==4.2.2
tushare==0.5.0
wcwidth==0.1.7
win-unicode-console==0.5
